requirements.txt
python-telegram-bot==20.3  # Укажите точную версию, используемую в проекте
feedparser>=6.0.11        # Последняя стабильная версия
requests>=2.32.0          # Для работы с HTTP-запросами
python-dotenv==1.0.0       # Для работы с .env файлами
